WebUI combine with auto run requirements.txt
